+++ demos/basic-migrations-raw/src/index.ts
@@ -1,10 +1,10 @@
-import { Db, Nullable } from "@ts-safeql-demos/shared/client";
+import { Db } from "@ts-safeql-demos/shared/client";
 import postgres from "postgres";
 
 export function check(client: Db) {
   const sql = postgres();
 
-  client.queryOne<{ id: number; post_id: Nullable<number>; body: Nullable<string> }>(sql`
+  client.queryOne<{ id: number; post_id: number | null; body: string | null }>(sql`
     SELECT * FROM comments
   `);
 }