+++ packages/generate/src/generate.test.ts
@@ -123,7 +123,7 @@ test("select with left join should return all cols from left join as nullable", 
         FROM caregiver
             LEFT JOIN caregiver_agency ON caregiver.id = caregiver_agency.caregiver_id
     `,
-    expected: `{ caregiver_id: number; assoc_id: Nullable<number>; }`,
+    expected: `{ caregiver_id: number; assoc_id: number | null; }`,
   });
 });